zinxV0.5

+++ zinx/znet/message.go
@@ -6,9 +6,9 @@ type Message struct {
 	Data    []byte // 消息内容
 }
 
-func NewMessgaPackage(id int32, data []byte) *Message {
+func NewMessgaPackage(id uint32, data []byte) *Message {
 	return &Message{
-		Id:      uint32(id),
+		Id:      id,
 		DataLen: uint32(len(data)),
 		Data:    data,
 	}
